#48 Create notification view, controller
