+++ practices/practice-one/src/scripts/controllers/notification.controller.ts
@@ -1,14 +1,19 @@
 import Notification from '../models/notification.model';
-import { showNotifications } from '../helpers/notification';
+import NotificationView from '../views/notification.view';
 
 
 export default class NotificationController {
-
+    private notificationView: NotificationView;
     constructor() {
+        this.notificationView = new NotificationView()
     }
 
+    /**
+     * - Handling display of messages on the screen
+     * @param notify 
+     */
     show(notify: Notification) {
         const time = notify.duringTime || 2500;
-        showNotifications(notify.message, time);
+        this.notificationView.show(notify.message, time);
     }
 }
