+++ practices/practice-one/src/scripts/views/todo.view.ts
@@ -52,6 +52,7 @@ export default class TodoView {
             };
             const todoItem = todoItemTemplate(param);
             this.todosElement.appendChild(todoItem);
+            this.addEventTodoItem(todoItem);
         });
         this.handleUpdateSizeTodo(this.hash);
     }
@@ -73,10 +74,53 @@ export default class TodoView {
         const todoItem = todoItemTemplate(data);
         if (this.hash !== TodoStates.COMPLETED) {
             this.todosElement.appendChild(todoItem);
+            this.addEventTodoItem(todoItem);
             this.handleUpdateSizeTodo(this.hash);
         }
         this.textElment.value = '';
         this.formGroupElemnt.classList.add('not-empty');
+    }
+
+    /**
+     * Add event todo item
+     * @pram liElement 
+     */
+    private addEventTodoItem(liElement: HTMLLIElement) {
+        const edit = liElement.querySelector('.form-control-input');
+        const btnComplete: HTMLButtonElement = liElement.querySelector('.btn-checkbox');
+        const inputUpdate: HTMLFormElement = liElement.querySelector('form');
+        const textElemnt: HTMLInputElement = inputUpdate.querySelector('input');
+        const btnDelete: HTMLButtonElement = liElement.querySelector('.btn-delete');
+        const pesudo: HTMLParagraphElement = liElement.querySelector('.pesudo .pesudo-value') as HTMLParagraphElement;
+        const timeUpdate: HTMLParagraphElement = liElement.querySelector('.pesudo-time-update') as HTMLParagraphElement;
+
+        btnDelete.addEventListener('click', () => {
+            this.handleDeletedTodo(liElement);
+        });
+
+        btnComplete.addEventListener('click', () => {
+            btnComplete.classList.toggle('checked');
+            liElement.classList.toggle('completed');
+            this.handleCompletedTodo(liElement);
+        });
+
+        pesudo.addEventListener('dblclick', () => {
+            edit.classList.add('edit');
+            textElemnt.focus();
+        });
+
+        inputUpdate.addEventListener('submit', (e) => {
+            e.preventDefault();
+            const result: Update = this.handleUpdateTodo(
+                pesudo,
+                textElemnt.value,
+                +liElement.getAttribute('data-item')
+            );
+            if (result.isUpdate) {
+                edit.classList.remove('edit');
+                timeUpdate.textContent = `UpdatedAt: ${result.time}`;
+            }
+        });
     }
 
     /**
